requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Machine Learning
tensorflow>=2.12.0
numpy>=1.24.0
h5py>=3.8.0
huggingface-hub>=0.20.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0

# PDF Generation
reportlab>=4.0.0

# Streamlit UI (alternative to FastAPI + React)
streamlit>=1.28.0
